requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
SQLAlchemy==2.0.32
aiosqlite==0.20.0
httpx==0.27.0
tenacity==8.5.0
python-jose==3.3.0
python-multipart==0.0.9
openai==1.40.3
orjson==3.10.7
